+++ .eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["airbnb-base", "plugin:@typescript-eslint/recommended"],
+  "extends": ["airbnb-base"],
   "rules": {
     "no-console": "off",
     "import/newline-after-import": "warn",
