+++ package.json
@@ -2,10 +2,9 @@
   "name": "crypto-price-tracker-cli",
   "version": "1.0.0",
   "description": "CLI tool that shows realtime cryptocurrency prices",
-  "main": "index.js",
+  "main": "src/app.js",
   "scripts": {
-    "start": "ts-node-dev src/app.ts",
-    "build": "tsc --build"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
@@ -18,26 +17,16 @@
   },
   "homepage": "https://github.com/RushDynamic/crypto-price-tracker-cli#readme",
   "devDependencies": {
-    "@tsconfig/node16": "^1.0.2",
-    "@types/commander": "^2.12.2",
-    "@types/inquirer": "^8.2.0",
-    "@types/node": "^17.0.21",
-    "@types/ws": "^8.5.2",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
     "eslint": "^8.10.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^16.1.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-prettier": "^4.0.0",
-    "prettier": "^2.5.1",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.6.2"
+    "prettier": "^2.5.1"
   },
   "dependencies": {
     "commander": "^9.0.0",
-    "inquirer": "^8.2.1",
     "ws": "^8.5.0"
-  }
+  },
+  "type": "module"
 }
