--- tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@tsconfig/node16/tsconfig.json",
-  "compilerOptions": {
-    "strict": true,
-    "target": "ES2016",
-    "module": "CommonJS",
-    "sourceMap": true,
-    "outDir": "dist",
-    "esModuleInterop": true
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
